(PUP-546) Fix problems with lookup after removal of hiera2
Moved lookup functionality to separate helper class.
