+++ lib/puppet/parser/functions/lookup.rb
@@ -33,11 +33,9 @@ specified, the default value is given to the block.
 
 The content of the options hash is:
 
-* `name` - The name to lookup. (Mutually exclusive with `first_found`)
+* `name` - The name or array of names to lookup (first found is returned)
 * `type` - The type to assert (a Type or a type specification in string form)
 * `default` - The default value if there was no value found (must comply with the data type)
-* `first_found` - An array of names to search, the value of the first found is used. (Mutually 
-  exclusive with `name`).
 * `accept_undef` - (default `false`) An `undef` result is accepted if this options is set to `true`.
 * `override` - a hash with map from names to values that are used instead of the underlying bindings. If the name
   is found here it wins. Defaults to an empty hash.
@@ -107,7 +105,7 @@ has the given data type (`String` in the example above):
 Using a lambda to process the looked up result - asserting that it starts with an upper case letter:
 
     # only with parser future
-    lookup('the_size', Integer[1,100) |$result| {
+    lookup('the_size', Integer[1,100]) |$result| {
       if $large_value_allowed and $result > 10
         { error 'Values larger than 10 are not allowed'}
       $result
@@ -116,21 +114,21 @@ Using a lambda to process the looked up result - asserting that it starts with a
 Including the name in the error
 
     # only with parser future
-    lookup('the_size', Integer[1,100) |$name, $result| {
+    lookup('the_size', Integer[1,100]) |$name, $result| {
       if $large_value_allowed and $result > 10
         { error 'The bound value for '${name}' can not be larger than 10 in this configuration'}
       $result
     }
 
-When using a block, the value it produces is also asserted against the given type, and it may not be 
+When using a block, the value it produces is also asserted against the given type, and it may not be
 `undef` unless the option `'accept_undef'` is `true`.
 
 All options work as the corresponding (direct) argument. The `first_found` option and
 `accept_undef` are however only available as options.
 
-Using the `first_found` option to return the first name that has a bound value:
+Using first_found semantics option to return the first name that has a bound value:
 
-    lookup({ first_found => ['apache::port', 'nginx::port'], type => 'Integer', default => 80})
+    lookup(['apache::port', 'nginx::port'], 'Integer', 80)
 
 If you want to make lookup return undef when no value was found instead of raising an error:
 
@@ -142,199 +140,5 @@ ENDHEREDOC
   unless Puppet[:binder] || Puppet[:parser] == 'future'
     raise Puppet::ParseError, "The lookup function is only available with settings --binder true, or --parser future" 
   end
-
-  def parse_lookup_args(args)
-    options = {}
-    pblock = if args[-1].respond_to?(:puppet_lambda)
-      args.pop
-    end
-
-    case args.size
-    when 1
-      # name, or all options
-      if args[ 0 ].is_a?(Hash)
-        options = to_symbolic_hash(args[ 0 ])
-      else
-        options[ :name ] = args[ 0 ]
-      end
-
-    when 2
-      # name and type, or name and options
-      if args[ 1 ].is_a?(Hash)
-        options = to_symbolic_hash(args[ 1 ])
-        options[:name] = args[ 0 ] # silently overwrite option with given name
-      else
-        options[:name] = args[ 0 ]
-        options[:type] = args[ 1 ]
-      end
-
-    when 3
-      # name, type, default (no options)
-      options[ :name ] = args[ 0 ]
-      options[ :type ] = args[ 1 ]
-      options[ :default ] = args[ 2 ]
-    else
-      raise Puppet::PareError, "The lookup function accepts 1-3 arguments, got #{args.size}"
-    end
-    options[:pblock] = pblock
-    options
-  end
-
-  def to_symbolic_hash(input)
-    names = [:name, :type, :first_found, :default, :accept_undef, :extra, :override]
-    options = {}
-    names.each {|n| options[n] = undef_as_nil(input[n.to_s] || input[n]) }
-    options
-  end
-
-  def type_mismatch(type_calculator, expected, got)
-    "has wrong type, expected #{type_calculator.string(expected)}, got #{type_calculator.string(got)}"
-  end
-
-  def fail(msg)
-    raise Puppet::ParseError, "Function lookup() " + msg
-  end
-
-  def fail_lookup(names)
-    name_part = if names.size == 1
-      "the name '#{names[0]}'"
-    else 
-      "any of the names ['" + names.join(', ') + "']"
-    end
-    fail("did not find a value for #{name_part}")
-  end
-
-  def validate_options(options, type_calculator)
-    type_parser = Puppet::Pops::Types::TypeParser.new
-    first_found_type = type_parser.parse('Array[String]')
-
-    if options[:name].nil? && options[:first_found].nil?
-      fail ("requires a name, or sequence of names in first_found. Neither was given.")
-    end
-
-    if options[:name] && options[:first_found]
-      fail("requires either a single name, or a sequence of names in first_found. Both were specified.")
-    end
-
-    if options[:name] && ! options[:name].is_a?(String)
-      t = type_calculator.infer(options[:name])
-      fail("name, expected String, got #{type_calculator.string(t)}")
-    end
-
-    if options[:first_found]
-      t = type_calculator.infer(options[:first_found])
-      if !type_calculator.assignable?(first_found_type, t)
-        fail("first_found #{type_mismatch(type_calculator, first_found_type, t)}")
-      end
-    end
-
-    # unless a type is already given (future case), parse the type (or default 'Data'), fails if invalid type is given
-    unless options[:type].is_a?(Puppet::Pops::Types::PAbstractType)
-      options[:type] = type_parser.parse(options[:type] || 'Data')
-    end
-
-    # default value must comply with the given type
-    if options[:default]
-      t = type_calculator.infer(options[:default])
-      if ! type_calculator.assignable?(options[:type], t)
-        fail("'default' value #{type_mismatch(type_calculator, options[:type], t)}")
-      end
-    end
-
-    if options[:extra] && !options[:extra].is_a?(Hash)
-      # do not perform inference here, it is enough to know that it is not a hash
-      fail("'extra' value must be a Hash, got #{options[:extra].class}")
-    end
-    options[:extra] = {} unless options[:extra]
-
-    if options[:override] && !options[:override].is_a?(Hash)
-      # do not perform inference here, it is enough to know that it is not a hash
-      fail("'override' value must be a Hash, got #{options[:extra].class}")
-    end
-    options[:override] = {} unless options[:override]
-
-  end
-
-  def nil_as_undef(x)
-    x.nil? ? :undef : x
-  end
-
-  def undef_as_nil(x)
-    is_nil_or_undef?(x) ? nil : x
-  end
-
-  def is_nil_or_undef?(x)
-    x.nil? || x == :undef
-  end
-
-  # This is used as a marker - a value that cannot (at least not easily) by mistake be found in
-  # hiera data.
-  #
-  class PrivateNotFoundMarker; end
-
-  def search_for(type, name, options)
-    # search in order, override, injector, hiera, then extra
-    if !(result = options[:override][name]).nil?
-      result
-    elsif !(result = compiler.injector.lookup(self, type, name)).nil?
-      result
-   else
-     result = self.function_hiera([name, PrivateNotFoundMarker])
-     if !result.nil? && result != PrivateNotFoundMarker
-       result
-     else
-       options[:extra][name]
-     end
-   end
-  end
-
-  # THE FUNCTION STARTS HERE
-
-  type_calculator = Puppet::Pops::Types::TypeCalculator.new
-  options = parse_lookup_args(args)
-  validate_options(options, type_calculator)
-  names = options[:name] ? [options[:name]] : options[:first_found]
-  type = options[:type]
-
-  result_with_name = names.reduce([]) do |memo, name|
-    break memo if !memo[1].nil?
-    [name, search_for(type, name, options)]
-  end
-
-  result = if result_with_name[1].nil?
-    # not found, use default (which may be nil), the default is already type checked
-    options[:default]
-  else
-    # injector.lookup is type-safe already do no need to type check the result
-    result_with_name[1]
-  end
-
-  result = if pblock = options[:pblock]
-    result2 = case pblock.parameter_count
-    when 1
-      pblock.call(self, nil_as_undef(result))
-    when 2
-      pblock.call(self, result_with_name[ 0 ], nil_as_undef(result))
-    else
-      pblock.call(self, result_with_name[ 0 ], nil_as_undef(result), nil_as_undef(options[ :default ]))
-    end
-
-    # if the given result was returned, there is not need to type-check it again
-    if !result2.equal?(result)
-      t = type_calculator.infer(undef_as_nil(result2))
-      if !type_calculator.assignable?(type, t)
-        fail "the value produced by the given code block #{type_mismatch(type_calculator, type, t)}"
-      end
-    end
-    result2
-  else
-    result
-  end
-
-  # Finally, the result if nil must be acceptable or an error is raised
-  if is_nil_or_undef?(result) && !options[:accept_undef]
-    fail_lookup(names)
-  else
-    nil_as_undef(result)
-  end
+  Puppet::Pops::Binder::Lookup.lookup(self, args)
 end
